# openapi.yaml
{ "openapi": "3.0.1","info": { "title": "HTTP SERVICE","version": "0.0.1 SNAPSHOT" },"servers": [ { "url": "http://localhost:8080","description": "Generated server url" } ],"tags": [ { "name": "Контроллер для работы с аккаунтом" },{ "name": "Контроллер для работы с ссылками" } ],"paths": { "/api/link/{id}": { "put": { "tags": [ "Контроллер для работы с ссылками" ],"summary": "Изменение пути в короткой ссылке","operationId": "changePath","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } },{ "name": "link","in": "query","required": true,"schema": { "type": "string" } },{ "name": "timestamp","in": "query","required": false,"schema": { "type": "integer","format": "int64","default": -1 } } ],"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/DataResponseLinkData" } } } } } },"delete": { "tags": [ "Контроллер для работы с ссылками" ],"summary": "Удаление ссылки","operationId": "deleteLink","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "200": { "description": "OK" } } } },"/api/link/{id}/recovery": { "put": { "tags": [ "Контроллер для работы с ссылками" ],"summary": "Восстановление ссылки","operationId": "recoveryLink","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/DataResponseLinkData" } } } } } } },"/api/link/create": { "post": { "tags": [ "Контроллер для работы с ссылками" ],"summary": "Создание короткой ссылки с учетом времени ее жизни","operationId": "registration","parameters": [ { "name": "link","in": "query","required": true,"schema": { "type": "string" } },{ "name": "timestamp","in": "query","required": false,"schema": { "type": "integer","format": "int64","default": -1 } } ],"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/DataResponseLinkData" } } } } } } },"/api/auth/register": { "post": { "tags": [ "Контроллер для работы с аккаунтом" ],"summary": "Регистрация","operationId": "registration_1","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } },"required": true },"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/DataResponseAuthData" } } } } } } },"/api/auth/login": { "post": { "tags": [ "Контроллер для работы с аккаунтом" ],"summary": "Авторизация","operationId": "login","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } },"required": true },"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/DataResponseAuthData" } } } } } } },"/api/link": { "get": { "tags": [ "Контроллер для работы с ссылками" ],"summary": "Получение списка коротких ссылок","operationId": "getAllLinks","parameters": [ { "name": "offset","in": "query","required": false,"schema": { "type": "integer","format": "int32","default": 0 } },{ "name": "itemPerPage","in": "query","required": false,"schema": { "type": "integer","format": "int32","default": 20 } } ],"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ListDataResponseLinkData" } } } } } } },"/api/auth/logout": { "get": { "tags": [ "Контроллер для работы с аккаунтом" ],"summary": "Выход из сессии","operationId": "getLogout","responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/DataResponseObject" } } } } } } } },"components": { "schemas": { "DataResponseLinkData": { "type": "object","properties": { "timestamp": { "type": "string","format": "date-time" },"data": { "$ref": "#/components/schemas/LinkData" } } },"LinkData": { "type": "object","properties": { "id": { "type": "integer","format": "int64" },"short_url": { "type": "string" },"view_count": { "type": "integer","format": "int32" },"unique_views": { "type": "object","additionalProperties": { "type": "integer","format": "int32" } } } },"RegisterRequest": { "type": "object","properties": { "email": { "type": "string" },"password": { "type": "string" } } },"AuthData": { "type": "object","properties": { "id": { "type": "integer","format": "int64" },"email": { "type": "string" } } },"DataResponseAuthData": { "type": "object","properties": { "timestamp": { "type": "string","format": "date-time" },"data": { "$ref": "#/components/schemas/AuthData" } } },"ListDataResponseLinkData": { "type": "object","properties": { "timestamp": { "type": "string","format": "date-time" },"total": { "type": "integer","format": "int32" },"offset": { "type": "integer","format": "int32" },"itemPerPage": { "type": "integer","format": "int32" },"data": { "type": "array","items": { "$ref": "#/components/schemas/LinkData" } } } },"DataResponseObject": { "type": "object","properties": { "timestamp": { "type": "string","format": "date-time" },"data": { "type": "object" } } } } } }